refactor: DTO 도메인 구조 및 필드 수정
