+++ spadeworker/src/main/java/com/study/spadeworker/domain/article/dto/article/CreateArticleDto.java
@@ -1,4 +1,4 @@
-package com.study.spadeworker.domain.article.dto;
+package com.study.spadeworker.domain.article.dto.article;
 
 import lombok.*;
 
@@ -8,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UpdateArticleDto {
+public class CreateArticleDto {
 
     @Getter
     @AllArgsConstructor
@@ -27,6 +27,9 @@ public class UpdateArticleDto {
         private String articleCategory;
 
         private List<String> hashtagList = new ArrayList<>();
+
+        @Setter
+        private Long boardId;
     }
 
     @Getter
