+++ spadeworker/src/main/java/com/study/spadeworker/domain/article/dto/articleComment/ArticleChildCommentDto.java
@@ -1,4 +1,4 @@
-package com.study.spadeworker.domain.article.dto;
+package com.study.spadeworker.domain.article.dto.articleComment;
 
 import com.study.spadeworker.domain.article.entity.ArticleComment;
 import com.study.spadeworker.domain.user.dto.UserAccountDto;
@@ -7,7 +7,7 @@ import lombok.Getter;
 import java.time.LocalDateTime;
 
 @Getter
-public class ChildArticleCommentDto {
+public class ArticleChildCommentDto {
 
     private Long id;
     private Long parentCommentId;
@@ -16,9 +16,10 @@ public class ChildArticleCommentDto {
     private int dislikesCount;
     private UserAccountDto recipient;
     private LocalDateTime createdAt;
+    private Long articleId;
     private UserAccountDto user;
 
-    private ChildArticleCommentDto(
+    private ArticleChildCommentDto(
                     Long id,
                     Long parentCommentId,
                     UserAccountDto recipient,
@@ -26,6 +27,7 @@ public class ChildArticleCommentDto {
                     int likesCount,
                     int dislikesCount,
                     LocalDateTime createdAt,
+                    Long articleId,
                     UserAccountDto userAccountDto
     ) {
         this.id = id;
@@ -35,15 +37,16 @@ public class ChildArticleCommentDto {
         this.likesCount = likesCount;
         this.dislikesCount = dislikesCount;
         this.createdAt = createdAt;
+        this.articleId = articleId;
         this.user = userAccountDto;
     }
 
-    public static ChildArticleCommentDto from(
+    public static ArticleChildCommentDto from(
             ArticleComment articleComment,
             UserAccountDto userAccountDto,
             UserAccountDto recipient
     ) {
-        return new ChildArticleCommentDto(
+        return new ArticleChildCommentDto(
                 articleComment.getId(),
                 articleComment.getParentComment().getId(),
                 recipient,
@@ -51,6 +54,7 @@ public class ChildArticleCommentDto {
                 articleComment.getLikesCount(),
                 articleComment.getDislikesCount(),
                 articleComment.getCreatedAt(),
+                articleComment.getArticle().getId(),
                 userAccountDto
         );
     }
